Split starterpack functionality.
