+++ helper.py
@@ -1,0 +1,2 @@
+def last(values):
+    return values[-1]
